[middleware/whitelist] an ability to build many instances of whitelist middleware
